Now user can add friends even by search

+++ app/Http/Controllers/SearchFriendsController.php
@@ -49,23 +49,49 @@ class SearchFriendsController extends Controller
     public function store(AddFriendsRequest $request)
     {
         $userId = Auth::id();
+
         $friendIds = $request->input('friendsIds', []);
+    
+        $searchTerm = $request->input('searchTerm', '');
+    
+        $totalFriendsIds = Friend::where('userId', $userId)->pluck('friendsId');
 
-        Friend::where('userId',$userId)->delete();
+       if (!empty($searchTerm)) {
+         $totalFriendsIds = $totalFriendsIds->toArray(); // Convert collection to array
+         $friendIds = $friendIds; // Assuming it's already an array
 
-        /*
-         * perdor if query string nuk eshte null te kapi
-         * te gjitha ID dhe nese ekziston ta fshij perndryshe ta shtoj
-         * */
+            // Merge the two arrays and remove duplicates
+          $totalArray = array_unique(array_merge($totalFriendsIds, $friendIds));
+            
+          Friend::where('userId', $userId)->delete();
 
-        foreach ($friendIds as $friendId) {
-            Friend::create([
-                'userId' => $userId,
-                'friendsId' => $friendId,
-                'status' => Friend::PENDING,
-            ]);
-        }
+            foreach ($totalArray as $friendId) {
+                Friend::create([
+                    'userId' => $userId,
+                    'friendsId' => $friendId,
+                    'status' => Friend::PENDING,
+                ]);
+            }            
+         } else {
+        
+            $totalFriendsIds = $totalFriendsIds->toArray(); // Convert collection to array
+            $friendIds = $friendIds; // Assuming it's already an array
 
+            // Merge the two arrays and remove duplicates
+            $totalArray = array_unique(array_merge($totalFriendsIds, $friendIds));
+
+            //dd($totalArray);
+            Friend::where('userId', $userId)->delete();
+
+            foreach ($friendIds as $friendId) {
+                Friend::create([
+                    'userId' => $userId,
+                    'friendsId' => $friendId,
+                    'status' => Friend::PENDING,
+                ]);
+         }
+    }
         return redirect()->back()->with('success', 'Selected friends saved successfully.');
     }
+    
 }
